[FIX] New face creation with wrong pairingEdges
In the loop that would close hull generation the pairing edges of the newly created edges would be incorrect. They have been fixed by adding the edge rather than the edge.pairingEdge to the list in the convexHullConstructHorizon function. Convexity checking was also incorrect because of checking the face with itself, messing up the normals of the faces.