[ADD] Option to destroy empty cells when the amount of empty cells are over a given amount
Currently the entire grid is reconstructed when the empty cells are cleared. This is necessary because removing empty cells from the existing cell list shifts all the indices of the remaining cells. These indices are stored in an unordered_map of ivec3 and int, which means that this list also needs to be edited to shift all the indices.